refactor: restructure code using modules design
