+++ config/gql-to-app-runtime-query/src/extractQueryDefinition/index.js
@@ -1,0 +1,2 @@
+const { extractQueryDefinition } = require('./extractQueryDefinition')
+module.exports.extractQueryDefinition = extractQueryDefinition
