+++ config/gql-to-app-runtime-query/src/extractResources/extractResources.js
