+++ config/gql-to-app-runtime-query/src/extractResources/index.js
@@ -1,0 +1,2 @@
+const { extractResources } = require('./extractResources')
+module.exports.extractResources = extractResources
