+++ config/gql-to-app-runtime-query/src/extractVariables/getGraphqlType.js
@@ -1,0 +1,4 @@
+module.exports.getGraphqlType = variableDefinition =>
+    variableDefinition.type.kind === 'NonNullType'
+        ? variableDefinition.type.type.name.value
+        : variableDefinition.type.name.value
