+++ config/gql-to-app-runtime-query/src/extractVariables/index.js
@@ -1,0 +1,2 @@
+const { extractVariables } = require('./extractVariables.js')
+module.exports.extractVariables = extractVariables
