+++ config/gql-to-app-runtime-query/src/handleMutation/index.js
@@ -1,0 +1,2 @@
+const { handleMutation } = require('./handleMutation')
+module.exports.handleMutation = handleMutation
