+++ config/gql-to-app-runtime-query/src/handleQuery/index.js
@@ -1,0 +1,2 @@
+const { handleQuery } = require('./handleQuery')
+module.exports.handleQuery = handleQuery
