+++ config/gql-to-app-runtime-query/src/resourceToObjectProperty/index.js
@@ -1,0 +1,2 @@
+const { resourceToObjectProperty } = require('./resourceToObjectProperty')
+module.exports.resourceToObjectProperty = resourceToObjectProperty
