+++ config/gql-to-app-runtime-query/src/resourceToObjectProperty/variablesToFunctionArgument.js
