--- config/gql-to-app-runtime-query/src/utils/pipe.js
@@ -1,4 +1,0 @@
-module.exports.pipe = (...fns) => (...args) => {
-    if (!fns.length) return
-    return fns.slice(1).reduce((acc, fn) => fn(acc), fns[0](...args))
-}
