+++ config/gql-to-app-runtime-query/src/validateHasSingleExpression/index.js
@@ -1,0 +1,2 @@
+const { validateHasSingleExpression } = require('./validateHasSingleExpression')
+module.exports.validateHasSingleExpression = validateHasSingleExpression
