+++ config/gql-to-app-runtime-query/src/validateHasSingleExpression/validateHasSingleExpression.js
@@ -1,0 +1,5 @@
+module.exports.validateHasSingleExpression = quasis => {
+    if (quasis.length > 1) {
+        throw new Error('"${}" Expressions are not allowed inside queries')
+    }
+}
