+++ config/gql-to-app-runtime-query/src/validateIsGqlTemplateExpression/index.js
@@ -1,0 +1,4 @@
+const {
+    validateIsGqlTemplateExpression,
+} = require('./validateIsGqlTemplateExpression')
+module.exports.validateIsGqlTemplateExpression = validateIsGqlTemplateExpression
