+++ config/gql-to-app-runtime-query/src/validateIsGqlTemplateExpression/validateIsGqlTemplateExpression.js
@@ -1,0 +1,4 @@
+module.exports.validateIsGqlTemplateExpression = node => {
+    const { tag } = node
+    return tag.name === 'gql'
+}
