+++ config/gql-to-app-runtime-query/src/validateVariableDefaultValues/index.js
@@ -1,0 +1,4 @@
+const {
+    validateVariableDefaultValues,
+} = require('./validateVariableDefaultValues')
+module.exports.validateVariableDefaultValues = validateVariableDefaultValues
